Move to plugins-common
